requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
pydantic
motor==2.2.0
aiofiles
python-multipart
uuid
cloudinary
python-decouple==3.3
